initial refactor moving everything UI to it's own module
