+++ molecularnodes/entities/__init__.py
@@ -1,23 +1,9 @@
 from . import molecule, trajectory
 from .base import EntityType, MolecularEntity
-from .density import MN_OT_Import_Map
 from .ensemble import CellPack, Ensemble, StarFile
-from .ensemble.ui import MN_OT_Import_Cell_Pack, MN_OT_Import_Star_File
 from .molecule import BCIF, CIF, PDB, SDF, Molecule
 from .molecule.pdb import PDB
 from .molecule.pdbx import BCIF, CIF
 from .molecule.sdf import SDF
-from .molecule.ui import fetch, load_local, parse
+from .molecule.io import fetch, load_local, parse
 from .trajectory import OXDNA, Trajectory
-from .trajectory.dna import MN_OT_Import_OxDNA_Trajectory
-
-CLASSES = (
-    [
-        MN_OT_Import_Cell_Pack,
-        MN_OT_Import_Map,
-        MN_OT_Import_OxDNA_Trajectory,
-        MN_OT_Import_Star_File,
-    ]
-    + trajectory.CLASSES
-    + molecule.CLASSES
-)
