+++ molecularnodes/entities/density/__init__.py
@@ -1,3 +1,3 @@
-from .ui import MN_OT_Import_Map, load
+from .io import load
 from .mrc import MRC
 from .base import Density
