+++ molecularnodes/entities/ensemble/__init__.py
@@ -1,4 +1,4 @@
 from .base import Ensemble
 from .cellpack import CellPack
 from .star import StarFile
-from .ui import load_cellpack, load_starfile
+from .io import load_cellpack, load_starfile
