+++ molecularnodes/entities/molecule/__init__.py
@@ -2,5 +2,4 @@ from .pdbx import BCIF, CIF
 from .pdb import PDB
 from .sdf import SDF
 from .base import Molecule
-
-from .ui import CLASSES
+from .io import fetch, load_local, parse
