+++ molecularnodes/entities/trajectory/__init__.py
@@ -1,7 +1,4 @@
 from . import selections
-from . import ui
-from .ui import load
+from .io import load, load_oxdna
 from .base import Trajectory
 from .dna import OXDNA
-
-CLASSES = ui.CLASSES
